# clinica_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clínica Social Cuidar - API System",
    "description": "\n**API Oficial de Gestão da Clínica Social Cuidar.**\n\n## Funcionalidades Principais:\n*   **Gestão de Pacientes**: Cadastro completo com criptografia de dados sensíveis (LGPD).\n*   **Prontuário Eletrônico**: Registro histórico de atendimentos.\n*   **Controle Financeiro**: Receitas e Despesas.\n*   **Auditoria**: Rastreabilidade total de ações.\n\n## Segurança e Governança:\n*   Todos os endpoints são protegidos (JWT).\n*   Rate Limiting ativo para prevenção de ataques.\n*   Dados sensíveis (CPF) são criptografados no banco (Fernet).\n        ",
    "contact": {
      "name": "Suporte Técnico - Vitor Moraes",
      "email": "suporte@clinicasocial.com.br"
    },
    "license": {
      "name": "Private License"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/attendance/my-appointments": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "Get My Appointments",
        "operationId": "get_my_appointments_api_v1_attendance_my_appointments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppointmentVolunteerRead"
                  },
                  "type": "array",
                  "title": "Response Get My Appointments Api V1 Attendance My Appointments Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/attendance/{appointment_id}/start": {
      "post": {
        "tags": [
          "attendance"
        ],
        "summary": "Start Attendance",
        "operationId": "start_attendance_api_v1_attendance__appointment_id__start_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/{appointment_id}/finish": {
      "post": {
        "tags": [
          "attendance"
        ],
        "summary": "Finish Attendance",
        "operationId": "finish_attendance_api_v1_attendance__appointment_id__finish_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/patient/{patient_id}/history": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "Get Patient History",
        "operationId": "get_patient_history_api_v1_attendance_patient__patient_id__history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalRecordRead"
                  },
                  "title": "Response Get Patient History Api V1 Attendance Patient  Patient Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/{appointment_id}/cancel": {
      "post": {
        "tags": [
          "attendance"
        ],
        "summary": "Cancel Attendance Start",
        "operationId": "cancel_attendance_start_api_v1_attendance__appointment_id__cancel_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/{appointment_id}/record": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "Get Appointment Record",
        "operationId": "get_appointment_record_api_v1_attendance__appointment_id__record_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentRead": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "volunteer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volunteer Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "scheduled"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price",
            "default": 0.0
          },
          "amount_paid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paid",
            "default": 0.0
          },
          "payment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Status",
            "default": "PENDING"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "date",
          "time",
          "id"
        ],
        "title": "AppointmentRead"
      },
...
      "AppointmentVolunteerRead": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "volunteer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volunteer Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "scheduled"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price",
            "default": 0.0
          },
          "amount_paid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paid",
            "default": 0.0
          },
          "payment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Status",
            "default": "PENDING"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "patient_name": {
            "type": "string",
            "title": "Patient Name"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "date",
          "time",
          "id",
          "patient_name"
        ],
        "title": "AppointmentVolunteerRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicalRecordCreate": {
        "properties": {
          "chief_complaint": {
            "type": "string",
            "title": "Chief Complaint"
          },
          "history": {
            "type": "string",
            "title": "History"
          },
          "procedures": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procedures"
          },
          "prescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescription"
          },
          "content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "chief_complaint",
          "history"
        ],
        "title": "MedicalRecordCreate"
      },
      "MedicalRecordRead": {
        "properties": {
          "chief_complaint": {
            "type": "string",
            "title": "Chief Complaint"
          },
          "history": {
            "type": "string",
            "title": "History"
          },
          "procedures": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procedures"
          },
          "prescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescription"
          },
          "content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "appointment_id": {
            "type": "string",
            "title": "Appointment Id"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "volunteer_id": {
            "type": "string",
            "title": "Volunteer Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "chief_complaint",
          "history",
          "id",
          "appointment_id",
          "patient_id",
          "volunteer_id",
          "created_at"
        ],
        "title": "MedicalRecordRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}